Zelda totk voice control for use with talon, vgamepad, ryujinx
