+++ roku_games/ryujinx/ryujinx_settings.talon
@@ -6,7 +6,7 @@ settings():
     key_wait = 16.0
 
     user.game_mode_disables_command_mode = true
-    user.game_xbox_left_stick_default_gear = 5
+    user.game_xbox_left_stick_default_gear = 3
     user.game_xbox_left_stick_gears = ".2 .4 .6 .8 1"
     user.game_xbox_right_stick_default_gear = 3
     user.game_xbox_right_stick_gears = ".4 .5 .6 .7 1"
@@ -17,9 +17,9 @@ settings():
     user.game_xbox_button_hold = 100
 
     # mouse movement for quick direction snaps
-    user.mouse_move_api = "windows"
-    user.game_calibrate_x_360 = 6000
-    user.game_calibrate_y_90 = 1500
+    #user.mouse_move_api = "windows"
+    #user.game_calibrate_x_360 = 6000
+    #user.game_calibrate_y_90 = 1500
 
     # exclude chars that have "hiss" sound
     user.drag_mode_exclude_chars = "giosvxz"
